+++ illuin_inject/scopes/per_lookup_scope.py
@@ -6,6 +6,5 @@ from .scope import Scope
 class PerLookupScope(Scope):
     """Provides a new instance every time."""
 
-    @classmethod
-    def get_scoped_provider(cls, inner_provider: Provider[InjectedT]) -> Provider[InjectedT]:
+    def get_scoped_provider(self, inner_provider: Provider[InjectedT]) -> Provider[InjectedT]:
         return inner_provider