+++ illuin_inject/scopes/thread_scope.py
@@ -7,6 +7,5 @@ from .thread_scoped_provider import ThreadScopedProvider
 class ThreadScope(Scope):
     """Always provides the same instance if called in the same thread, creates a new one if not."""
 
-    @classmethod
-    def get_scoped_provider(cls, inner_provider: Provider[InjectedT]) -> Provider[InjectedT]:
+    def get_scoped_provider(self, inner_provider: Provider[InjectedT]) -> Provider[InjectedT]:
         return ThreadScopedProvider(inner_provider)